+++ src/Adafruit_EPD.h
@@ -64,11 +64,11 @@ typedef enum {
 /**************************************************************************/
 class Adafruit_EPD : public Adafruit_GFX {
 public:
-  Adafruit_EPD(int width, int height, int8_t SID, int8_t SCLK, int8_t DC,
-               int8_t RST, int8_t CS, int8_t SRCS, int8_t MISO,
-               int8_t BUSY = -1);
-  Adafruit_EPD(int width, int height, int8_t DC, int8_t RST, int8_t CS,
-               int8_t SRCS, int8_t BUSY = -1, SPIClass *spi = &SPI);
+  Adafruit_EPD(int width, int height, int16_t SID, int16_t SCLK, int16_t DC,
+               int16_t RST, int16_t CS, int16_t SRCS, int16_t MISO,
+               int16_t BUSY = -1);
+  Adafruit_EPD(int width, int height, int16_t DC, int16_t RST, int16_t CS,
+               int16_t SRCS, int16_t BUSY = -1, SPIClass *spi = &SPI);
   ~Adafruit_EPD();
 
   void begin(bool reset = true);
@@ -132,7 +132,7 @@ protected:
   virtual void powerDown(void) = 0;
   void hardwareReset(void);
 
-  int8_t _dc_pin,                     ///< data/command pin
+  int16_t _dc_pin,                    ///< data/command pin
       _reset_pin,                     ///< reset pin
       _cs_pin,                        ///< chip select pin
       _busy_pin;                      ///< busy pin
