+++ src/drivers/Adafruit_ACeP.h
@@ -35,11 +35,11 @@
 /**************************************************************************/
 class Adafruit_ACEP : public Adafruit_EPD {
 public:
-  Adafruit_ACEP(int width, int height, int8_t SID, int8_t SCLK, int8_t DC,
-                int8_t RST, int8_t CS, int8_t SRCS, int8_t MISO,
-                int8_t BUSY = -1);
-  Adafruit_ACEP(int width, int height, int8_t DC, int8_t RST, int8_t CS,
-                int8_t SRCS, int8_t BUSY = -1, SPIClass *spi = &SPI);
+  Adafruit_ACEP(int width, int height, int16_t SID, int16_t SCLK, int16_t DC,
+                int16_t RST, int16_t CS, int16_t SRCS, int16_t MISO,
+                int16_t BUSY = -1);
+  Adafruit_ACEP(int width, int height, int16_t DC, int16_t RST, int16_t CS,
+                int16_t SRCS, int16_t BUSY = -1, SPIClass *spi = &SPI);
 
   void begin(bool reset = true);
   void powerUp();
