+++ src/panels/ThinkInk_290_Mono_BN.h
@@ -6,12 +6,13 @@
 
 class ThinkInk_290_Mono_BN : public Adafruit_SSD1680 {
 public:
-  ThinkInk_290_Mono_BN(int8_t SID, int8_t SCLK, int8_t DC, int8_t RST,
-                       int8_t CS, int8_t SRCS, int8_t MISO, int8_t BUSY = -1)
+  ThinkInk_290_Mono_BN(int16_t SID, int16_t SCLK, int16_t DC, int16_t RST,
+                       int16_t CS, int16_t SRCS, int16_t MISO,
+                       int16_t BUSY = -1)
       : Adafruit_SSD1680(296, 128, SID, SCLK, DC, RST, CS, SRCS, MISO, BUSY){};
 
-  ThinkInk_290_Mono_BN(int8_t DC, int8_t RST, int8_t CS, int8_t SRCS,
-                       int8_t BUSY = -1, SPIClass *spi = &SPI)
+  ThinkInk_290_Mono_BN(int16_t DC, int16_t RST, int16_t CS, int16_t SRCS,
+                       int16_t BUSY = -1, SPIClass *spi = &SPI)
       : Adafruit_SSD1680(296, 128, DC, RST, CS, SRCS, BUSY, spi){};
 
   void begin(thinkinkmode_t mode = THINKINK_MONO) {
